Add k-corrected Hunter–Schmidt estimator
This estimator is like the HS estimator in the τ = var_observed - var_expected, but it computes the weighted observed variance of yi using the unbiased estimator (dividing by k-1 instead of k); see https://doi.org/10.1177/1094428117741966. This estimator is similar to the DL estimator, but computes the denominator as sum(wi) * (k-p) / k rather than trP; that it, it computes the denominator assuming equal weights.

This estimator is used fairly often when the Hunter-Schmidt method is applied. It's the default estimator used in psychmeta. With this addition, we would be able to perfectly reproduce psychmeta results using metafor with the correct specification of vi and weights.